+++ SportsHub/UnitTests/Services/TestAuthenticationService.cs
@@ -21,12 +21,15 @@ public class TestAuthenticationService
     [Fact]
     public async Task Authenticate_WithUsername_ReturnUser()
     {
-        var givenUser = UserLoginDTOMockData.GetUserWithUsername();
+        //Arrange
+        var givenUser = UserMockData.GetUserWithUsername();
         var user = UserMockData.GetUser();
         _userService.Setup(service => service.GetByUsernameAsync(givenUser.UsernameOrEmail)).Returns(Task.FromResult<User?>(user));
 
+        //Act
          var result = await _authentication.Authenticate(givenUser); 
-        
+       
+         //Assert
         Assert.Equal(result.Username, user.Username );
         Assert.Equal(result.Password, user.Password);
         
@@ -35,36 +38,44 @@ public class TestAuthenticationService
     [Fact]
     public async Task Authenticate_WithUsername_ReturnNull()
     {
-        var givenUser = UserLoginDTOMockData.GetUserWithUsername();
+        //Arrange
+        var givenUser = UserMockData.GetUserWithUsername();
         _userService.Setup(service => service.GetByUsernameAsync(givenUser.UsernameOrEmail)).Returns(Task.FromResult<User?>(null));
 
+        //Act
         var result = await _authentication.Authenticate(givenUser);
-        
+      
+        //Assert
         Assert.Null(result);
     }
     
     [Fact]
     public async Task Authenticate_WithEmail_ReturnUser()
     {
-        var givenUser = UserLoginDTOMockData.GetUserWithEmail();
+        //Arrange
+        var givenUser = UserMockData.GetUserWithEmail();
         var user = UserMockData.GetUser();
         _userService.Setup(service => service.GetByEmailAsync(givenUser.UsernameOrEmail)).Returns(Task.FromResult<User?>(user));
 
+        //Act
         var result = await _authentication.Authenticate(givenUser); 
-        
+      
+        //Assert
         Assert.Equal(result.Email, user.Email );
         Assert.Equal(result.Password, user.Password);
-        
     }
     
     [Fact]
     public async Task Authenticate_WithEmail_ReturnNull()
     {
-        var givenUser = UserLoginDTOMockData.GetUserWithEmail();
+        //Arrange
+        var givenUser = UserMockData.GetUserWithEmail();
         _userService.Setup(service => service.GetByUsernameAsync(givenUser.UsernameOrEmail)).Returns(Task.FromResult<User?>(null));
 
+        //Act
         var result = await _authentication.Authenticate(givenUser);
-        
+      
+        //Assert
         Assert.Null(result);
     }
 }
