+++ SportsHub/UnitTests/Services/TestUserService.cs
@@ -19,7 +19,7 @@ public class TestUserService : IDisposable
             .Options;
         _context = new ApplicationDbContext(options);
         _context.Database.EnsureCreated();
-        _context.Users.AddRange(MockData.UsersMockData.GetUsers());
+        _context.Users.AddRange(MockData.UserMockData.GetUsers());
         _unitOfWork = new UnitOfWork(_context);
         _unitOfWork.SaveChangesAsync();
     }
@@ -29,10 +29,13 @@ public class TestUserService : IDisposable
     [InlineData("seconduser@mail.com")]
     public async Task GetByEmailAsync_WithCorrectEmail_ReturnsCorrectUser(string email)
     {
+        //Arrange
         var sut = new UserService(_unitOfWork);
 
+        //Act
         var result = await sut.GetByEmailAsync(email);
 
+        //Assert
         Assert.Equal(email, result.Email);
     }
 
@@ -41,10 +44,13 @@ public class TestUserService : IDisposable
     [InlineData(null)]
     public async Task GetByEmailAsync_WithIncorrectEmail_ReturnsNull(string email)
     {
+        //Arrange
         var sut = new UserService(_unitOfWork);
 
+        //Act
         var result = await sut.GetByEmailAsync(email);
 
+        //Assert
         Assert.Null(result);
     }
 
@@ -53,10 +59,13 @@ public class TestUserService : IDisposable
     [InlineData("niki")]
     public async Task GetByUsernameAsync_WithCorrectUsername_ReturnsCorrectUser(string userName)
     {
+        //Arrange
         var sut = new UserService(_unitOfWork);
 
+        //Act
         var result = await sut.GetByUsernameAsync(userName);
 
+        //Assert
         Assert.Equal(userName, result.Username);
     }
 
@@ -64,10 +73,13 @@ public class TestUserService : IDisposable
     [InlineData("peshkata")]
     public async Task GetByUsernameAsync_WithIncorrectUsername_ReturnsNull(string userName)
     {
+        //Arrange
         var sut = new UserService(_unitOfWork);
 
+        //Act
         var result = await sut.GetByUsernameAsync(userName);
 
+        //Assert
         Assert.Null(result);
     }
